+++ README.md
@@ -88,6 +88,7 @@ end
 | deno        | https://deno.land/                                                          |
 | diagnosticls| https://github.com/iamcco/diagnostic-languageserver                         |
 | efm         | https://github.com/mattn/efm-langserver                                     |
+| jedi        | https://github.com/pappasam/jedi-language-server                            |
 | rome        | https://rome.tools/                                                         |
 
 Note: css, json and html language servers are pulled directly from the latest VSCode release, instead of using the outdated versions provided by e.g. `npm install vscode-html-languageserver-bin`.
