+++ lua/lspinstall/servers.lua
@@ -11,7 +11,7 @@ local servers = {
   ["go"] = require'lspinstall/servers/go',
   ["graphql"] = require'lspinstall/servers/graphql',
   ["html"] = require'lspinstall/servers/html',
-  ["jedi_language_server"] = require'lspinstall/servers/jedi_language_server',
+  ["jedi"] = require'lspinstall/servers/jedi',
   ["json"] = require'lspinstall/servers/json',
   ["latex"] = require'lspinstall/servers/latex',
   ["lua"] = require'lspinstall/servers/lua',
